+++ tests/django_app/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class DjangoAppConfig(AppConfig):
+    name = "tests.django_app"
+    label = "django_app"